requirements.txt
google-generativeai>=0.8.0
Pillow>=10.0.0
pydantic>=2.5.0
python-dotenv>=1.0.0
streamlit>=1.28.0
requests>=2.31.0 